+++ F-Droid/src/org/fdroid/fdroid/FDroid.java
@@ -137,6 +137,7 @@ public class FDroid extends AppCompatActivity implements SearchView.OnQueryTextL
     @Override
     protected void onResume() {
         super.onResume();
+        FDroidApp.checkStartTor(this);
         // AppDetails and RepoDetailsActivity set different NFC actions, so reset here
         NfcHelper.setAndroidBeam(this, getApplication().getPackageName());
         checkForAddRepoIntent(getIntent());